+++ app/TodoList.tsx
@@ -1,6 +1,6 @@
 import { observer } from 'mobx-react'
 import React, { Component } from 'react';
-import { FlatList, View, Text, AppRegistry, StyleSheet, ScrollView } from 'react-native';
+import { FlatList, AppRegistry, ScrollView } from 'react-native';
 import { todoStore } from './todoStore';
 import TodoItem from './TodoItem';
 
@@ -13,7 +13,7 @@ export class TodoList extends Component {
                 <FlatList data={todoStore.todo}
 
                 renderItem={
-                    ({item}) => <TodoItem title={item.title} />
+                    ({item}) => <TodoItem todo={item} />
                 }
                 keyExtractor={(item, index) => index.toString()}