+++ app/todo.ts
@@ -1,6 +1,9 @@
+import { observable } from "mobx";
+
 export class Todo {
-    title: string;
-    completed: boolean;
+    @observable title: string;
+    @observable completed: boolean;
+
     constructor(title: string, completed: boolean) {
         this.title = title;
         this.completed = completed;