+++ CHANGELOG.md
@@ -1,3 +1,10 @@
- - Added `TestScreen`
- - Reworked `TodoItem`
- - Simplified `TodoList`
+# Introduced AsyncStorage
+
+ - Created getData and setData functions in `TodoStore`
+ - Created component `InputScreen` and removed `<TodoInput/>` from `TodoScreen`
+ - Made `TodoItem` be able to toggle "completed" switch
+ - Cleaned up some unused imports
+ - Introduced trimming of input strings
+
+# TODO:
+ - Store and display descriptions of todo items (replace the placeholder)